Edmx Update,Add House - Complete

+++ LYSAdmin.Web/LYSAdmin.Model/Block.cs
@@ -9,6 +9,7 @@ namespace LYSAdmin.Model
 {
     public class Block
     {
+
         public int BlockID { get; set; }
         [Required]
         public string BlockName { get; set; }
@@ -20,7 +21,7 @@ namespace LYSAdmin.Model
         public Nullable<bool> isDeleted { get; set; }
         public Nullable<int> DeletedBy { get; set; }
         public Nullable<System.DateTime> DeletedOn { get; set; }
-        public Nullable<int> CreatedBy { get; set; }
+        public Nullable<long> CreatedBy { get; set; }
         public bool IsDefault { get; set; }
 
         public virtual Apartment Apartment { get; set; }
