+++ LYSAdmin.Web/LYSAdmin.Model/House.cs
@@ -8,15 +8,16 @@ namespace LYSAdmin.Model
 {
     public class House
     {
+
         public int HouseID { get; set; }
         public string HouseName { get; set; }
         public Nullable<bool> Status { get; set; }
         public Nullable<System.DateTime> CreatedOn { get; set; }
         public Nullable<System.DateTime> LastUpdatedOn { get; set; }
         public Nullable<bool> isDeleted { get; set; }
-        public Nullable<int> DeletedBy { get; set; }
+        public Nullable<long> DeletedBy { get; set; }
         public Nullable<System.DateTime> DeletedOn { get; set; }
-        public Nullable<int> CreatedBy { get; set; }
+        public Nullable<long> CreatedBy { get; set; }
         public string DisplayName { get; set; }
         public Nullable<int> Gender { get; set; }
         public Nullable<int> NoOfBathrooms { get; set; }
